elevenlabs_based/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --python-version 3.10
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
click==8.3.3
    # via uvicorn
exceptiongroup==1.3.1
    # via anyio
fastapi==0.115.5
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.13
    # via
    #   anyio
    #   httpx
pydantic==2.13.4
    # via fastapi
pydantic-core==2.46.4
    # via pydantic
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.3
    # via uvicorn
starlette==0.41.3
    # via fastapi
truststore==0.10.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.32.1
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0
    # via
    #   -r requirements.in
    #   uvicorn
